requirements.txt
boto3>=1.34.0
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
cachetools>=5.3.0
python-dotenv>=1.0.0
